fix: specDone produces result object with 'suite'
KarmaReporter.specDone was creating a result object with 'suites'
instead of 'suite', which was crashing Karma.

Additionally, the spec description was being pushed into the suite
array, causing Karma to show spec description twice in the spec name.

Suite name was also wrongly sliced, as jasmine appends an additional
period to fullName.